Refactor(plugins): Move jinja filter code for arista.avd.range_expand to PyAVD
