+++ python-avd/pyavd/_eos_designs/eos_designs_facts/uplinks.py
@@ -8,8 +8,8 @@ from functools import cached_property
 from typing import TYPE_CHECKING
 
 from ...j2filters.list_compress import list_compress
+from ...j2filters.range_expand import range_expand
 from ...vendor.errors import AristaAvdError
-from ...vendor.j2.filter.range_expand import range_expand
 from ...vendor.utils import append_if_not_duplicate, get, unique
 
 if TYPE_CHECKING:
