+++ python-avd/pyavd/_eos_designs/eos_designs_facts/vlans.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING
 
 from ...j2filters.convert_dicts import convert_dicts
 from ...j2filters.list_compress import list_compress
-from ...vendor.j2.filter.range_expand import range_expand
+from ...j2filters.range_expand import range_expand
 from ...vendor.utils import get
 
 if TYPE_CHECKING:
