+++ python-avd/pyavd/_eos_designs/shared_utils/cv_topology.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from ...vendor.j2.filter.range_expand import range_expand
+from ...j2filters.range_expand import range_expand
 from ...vendor.utils import get, get_item
 
 if TYPE_CHECKING:
