+++ python-avd/pyavd/_eos_designs/shared_utils/filtered_tenants.py
@@ -9,6 +9,7 @@ from typing import TYPE_CHECKING
 from ..._utils.merge import merge
 from ...j2filters.convert_dicts import convert_dicts
 from ...j2filters.natural_sort import natural_sort
+from ...j2filters.range_expand import range_expand
 from ...vendor.errors import AristaAvdError, AristaAvdMissingVariableError
 from ...vendor.j2.filter.range_expand import range_expand
 from ...vendor.utils import default, get, get_item, unique
