+++ python-avd/pyavd/_eos_designs/structured_config/underlay/utils.py
@@ -7,8 +7,8 @@ from functools import cached_property
 from typing import TYPE_CHECKING
 
 from ....j2filters.natural_sort import natural_sort
+from ....j2filters.range_expand import range_expand
 from ....vendor.errors import AristaAvdError
-from ....vendor.j2.filter.range_expand import range_expand
 from ....vendor.strip_empties import strip_empties_from_dict
 from ....vendor.utils import default, get, get_item
 from ...interface_descriptions import InterfaceDescriptionData
