+++ python-avd/pyavd/_eos_designs/structured_config/underlay/vlans.py
@@ -7,7 +7,7 @@ from functools import cached_property
 from typing import TYPE_CHECKING
 
 from ....j2filters.natural_sort import natural_sort
-from ....vendor.j2.filter.range_expand import range_expand
+from ....j2filters.range_expand import range_expand
 from ....vendor.utils import get, get_item
 from .utils import UtilsMixin
