+++ python-avd/pyavd/templater.py
@@ -80,7 +80,8 @@ class Templar:
         from .j2filters.natural_sort import natural_sort
         from .j2filters.snmp_hash import snmp_hash
         from .j2tests.contains import contains
-        from .vendor.j2.filter.range_expand import range_expand
+        from .vendor.j2.filter.decrypt import decrypt
+        from .vendor.j2.filter.encrypt import encrypt
         from .vendor.j2.test.defined import defined
 
         # pylint: enable=import-outside-toplevel
