+++ python-avd/tests/pyavd/j2filters/test_range_expand.py
@@ -1,39 +1,40 @@
 # Copyright (c) 2023-2024 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
+# Copyright (c) 2023-2024 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
 from __future__ import absolute_import, division, print_function
 
-__metaclass__ = type
-
 import pytest
+from pyavd.j2filters.range_expand import range_expand
 
-from ansible_collections.arista.avd.plugins.filter.range_expand import AnsibleFilterError, FilterModule, range_expand
 
 RANGE_TO_EXPAND_INVALID_VALUES = [
-    pytest.param(True, AnsibleFilterError, "value must be of type list or str, got <class 'bool'>", id="Wrong input type - bool"),
-    pytest.param({"key": "value"}, AnsibleFilterError, "value must be of type list or str, got <class 'dict'>", id="Wrong input type - dict"),
-    pytest.param(33, AnsibleFilterError, "", id="Wrong input type - int"),
+    pytest.param(True, TypeError, "value must be of type list or str, got <class 'bool'>", id="Wrong input type - bool"),
+    pytest.param({"key": "value"}, TypeError, "value must be of type list or str, got <class 'dict'>", id="Wrong input type - dict"),
+    pytest.param(33, TypeError, "", id="Wrong input type - int"),
     pytest.param(
         "Ethernet4-2",
-        AnsibleFilterError,
+        Exception,
         "Range Ethernet4-2 could not be expanded because the first interface 4 is larger than last interface 2 in the range.",
         id="Wrong interface range",
     ),
     pytest.param(
         "Ethernet1,51-3/2",
-        AnsibleFilterError,
+        Exception,
         "Range 51-3/2 could not be expanded because the first module 51 is larger than last module 3 in the range.",
         id="Wrong module range",
     ),
     pytest.param(
         "Ethernet1.42-21",
-        AnsibleFilterError,
+        Exception,
         "Range Ethernet1.42-21 could not be expanded because the first subinterface 42 is larger than last subinterface 21 in the range.",
         id="Wrong subinterface range",
     ),
     pytest.param(
         "Ethernet4/2-1/4",
-        AnsibleFilterError,
+        Exception,
         "Range Ethernet4/2-1/4 could not be expanded because the first interface 2 is larger than last interface 1 in the range.",
         id="Wrong parent interface range",
     ),
@@ -79,8 +80,6 @@ EXPECTED_RESULT_VALID_VALUES = [
     ["1.0", "1.1", "2.0", "2.1"],
 ]
 
-f = FilterModule()
-
 
 class TestRangeExpandFilter:
     @pytest.mark.parametrize("input_value, expected_raise, expected_raise_message", RANGE_TO_EXPAND_INVALID_VALUES)
@@ -92,8 +91,3 @@ class TestRangeExpandFilter:
     def test_range_expand_valid(self, RANGE_TO_EXPAND_VALID):
         resp = range_expand(RANGE_TO_EXPAND_VALID)
         assert resp in EXPECTED_RESULT_VALID_VALUES
-
-    def test_range_expand_filter(self):
-        resp = f.filters()
-        assert isinstance(resp, dict)
-        assert "range_expand" in resp.keys()
